building all qso,star,elg,lrg priors for OSU workshop
